Removes Tile from cache after a period of inactivity.
With the current architecture, a Tile is loaded from disk on
demand (i.e. when a pair of coordinates happen to fall within its
boundaries and it's not yet in memory). This choice was made to keep
latency (mostly) low without having to load the entire dataset in memory
-- a machine with that much RAM would be too costly, and our workload is
usually confined to one or two degrees ineither direction).

The counterpoint is that, loading a priori or on demand, you can still
crash the service by having too little memory. What this commit
introduces is the ability to keep data in memory for a while, renewing
its stay as it's accessed. After some inactivity (six hours by default),
the tile is thrown away.

Service had a tile cache, now extracted to a real structure -- an actor
that manages the inactivity periods, loads data as needed and unloads
them when they're stale.