Uses WebMercator to interpolate route points.
The goal was to solve the problem without relying on geodesic
computations. Despite the possibility of error in long distances, using
WebMercator to interpolate the route and converting it back to Geo when
done has given us precise enough results. The generative tests take care
of verifying the new, interpolated route is within acceptable distance
of the original one.